Move validation helpers to shared namespace
