+++ source/ProcessManager.Orchestrations/Processes/BRS_026_028/Shared/BusinessValidation/EnergySupplierIsOnlyAllowedToRequestOwnDataHelper.cs
@@ -16,7 +16,7 @@ using Energinet.DataHub.ProcessManager.Components.BusinessValidation;
 using Energinet.DataHub.ProcessManager.Components.BusinessValidation.Helpers;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026_028.BRS_026.V1.BusinessValidation.Helpers;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026_028.Shared.BusinessValidation;
 
 public static class EnergySupplierIsOnlyAllowedToRequestOwnDataHelper
 {
