+++ source/ProcessManager.Orchestrations/Processes/BRS_026_028/Shared/BusinessValidation/GridAreaValidationHelper.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.ProcessManager.Components.BusinessValidation.GridAreaOwner;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026_028.BRS_026.V1.BusinessValidation.Helpers;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026_028.Shared.BusinessValidation;
 
 public static class GridAreaValidationHelper
 {
