+++ source/ProcessManager.Orchestrations/Processes/BRS_026_028/Shared/BusinessValidation/SettlementVersionValidationHelper.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026_028.BRS_026.V1.BusinessValidation.Helpers;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026_028.Shared.BusinessValidation;
 
 public static class SettlementVersionValidationHelper
 {
